adicionado DAO estilo hibernate e conexao com banco aws, adicionado rota de pegar produto por id

+++ src/main/java/br/com/dificuldadezero/management/dao/ProductHibernateDAO.java
@@ -81,7 +81,7 @@ public class ProductHibernateDAO {
                    }
          }
   
-         public void removeById(final int id) {
+         public void removeById(final long id) {
                    try {
                             Product product = getById(id);
                             remove(product);
